refactor(User): move component to its own directory
