+++ src/components/appbar/AppBar.tsx
@@ -9,7 +9,7 @@ import {
 } from '@mui/material'
 import ChangeTheme from '~components/theme/ChangeTheme'
 import { LogoutOutlined } from '@mui/icons-material'
-import UserMenu from '~components/menu/User'
+import UserMenu from '~/components/menu/User/User'
 import { useAppDispatch } from '~/redux/hooks'
 import { logout } from '~/redux/features/login/loginSlice'
